use color function for signs
