+++ lmfdb/lfunctions/LfunctionPlot.py
@@ -7,7 +7,9 @@ from lmfdb.modular_forms.elliptic_modular_forms.backend.web_newforms import WebN
 from lmfdb.modular_forms.elliptic_modular_forms.backend.web_modform_space import WebModFormSpace
 from lmfdb.characters.ListCharacters import get_character_modulus
 from lmfdb.lfunctions import logger
-from sage.all import prod, CC
+from sage.all import prod
+from lmfdb.utils import signtocolour
+
 
 ###############################################################################
 # Maass form for GL(n) n>2
@@ -716,21 +718,6 @@ def paintCSHoloTMP(width, height, xMax, yMax, xfactor, yfactor, ticlength):
     return(xmlText)
 
 ##================================================
-#+++++++++++++++++++++++++++++++++++++++++++++++++
-##
-##================================================
-
-
-def signtocolour(sign):
-    import cmath
-    argument = cmath.phase(CC(str(sign)))
-    r = int(255.0 * (math.cos((1.0 * math.pi / 3.0) - (argument / 2.0))) ** 2)
-    g = int(255.0 * (math.cos((2.0 * math.pi / 3.0) - (argument / 2.0))) ** 2)
-    b = int(255.0 * (math.cos(argument / 2.0)) ** 2)
-    return("rgb(" + str(r) + "," + str(g) + "," + str(b) + ")")
-
-#=====================
-
 
 ###############################################################################
 # Dirichlet characters
