+++ lmfdb/lfunctions/main.py
@@ -23,6 +23,7 @@ import LfunctionDatabase
 from lmfdb import base
 from pymongo import ASCENDING
 from lmfdb.modular_forms.maass_forms.maass_waveforms.views.mwf_plot import paintSvgMaass
+from lmfdb.utils import signtocolour, rgbtohex
 
 def get_degree(degree_string):
     if not re.match('degree[0-9]+',degree_string):
@@ -104,6 +105,8 @@ def l_function_maass_browse_page():
     info = {"bread": get_bread(2, [("MaassForm", url_for('.l_function_maass_browse_page'))])}
     info["contents"] = [processMaassNavigation()]
     info["gl2spectrum0"] = [paintSvgMaass(1, 10, 0, 10, L="/L")]
+    info["colorminus1"] = rgbtohex(signtocolour(-1))
+    info["colorplus1"] = rgbtohex(signtocolour(1))
     return render_template("MaassformGL2.html", title='L-functions of GL(2) Maass Forms of weight 0', **info)
