+++ lmfdb/lfunctions/templates/MaassformGL2.html
@@ -27,9 +27,9 @@ when $f$ is odd (or even).
 <div>
 The dots in the plot correspond to L-functions with trivial character and
 $(R, N)$ as in the functional equation.
-Each eigenvalue is colored <font color="red"><b>red</b></font> 
-or <font color="blue"><b>blue</b></font> depending on whether
-the corresponding Maass cusp form is <font color="red"><b>even</b></font> or <font color="blue"><b>odd</b></font>.
+The color shows whether the functional equation has <b><font color="{{colorplus1}}">sign&nbsp;+1</font></b>
+or
+<b><font color="{{colorminus1}}">sign&nbsp;-1</font></b>.
 These have been found by a computer search.
 Click on any of the dots for details about the L-function.
 </div>
