+++ lmfdb/modular_forms/maass_forms/maass_waveforms/views/templates/mwf_browse_graph.html
@@ -3,7 +3,13 @@
 {% block content %}
 <h1>Graph of eigenvalues of Maass forms of levels {{min_level}} to {{max_level}} with ${{min_R}} \leq R\leq {{max_R}}$</h1>
 <p>
-    {{KNOWL_INC('mf.maass.mwf.browsegraph')}}
+   The horizontal axis is the {{KNOWL('mf.maass.mwf.spectralparameter', title='spectral parameter')}} $R$ with the {{KNOWL('mf.maass.mwf.eigenvalue', title='Laplace eigenvalue')}}
+satisying $\lambda=1/4+R²$. The vertical axis is the {{KNOWL('mf.maass.mwf.level', title='level')}} $N$. Each point
+corresponds to a Maass form of {{KNOWL('mf.maass.mwf.weight', title='weight')}} 0 and trivial character on $\Gamma_0(N)$
+with the color showing whether the {{KNOWL('mf.maass.mwf.symmetry', title='symmetry')}} is <b><font color="{{coloreven}}">even</font></b>
+or <b><font color="{{colorodd}}">odd</font></b>. Clicking on a dot takes you
+to the homepage of the Maass form. For $N>100$ there are only results for prime level.
+
 </p>
 <p>
 Examples of some ranges with complete data:
